+++ CHANGELOG.md
@@ -88,3 +88,4 @@
 - add PrePersist model event that adds createdAt and updatedAt date
 - add PreUpdate model event that update updatedAt date
 - base model extraction of Long ID, Date createdAt and Date updatedAt columns
+- add UUID unique identifier to base model